% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_submit.R
\name{apply_submit}
\alias{apply_submit}
\title{Perform a try to submit the filled form
contains validation und writing the datas if validation is valid}
\usage{
apply_submit(input, output, output_dir)
}
\arguments{
\item{input}{inputs passed-through server function}

\item{output}{output passed-through server function}

\item{output_dir}{directory path, where all orders are in}
}
\description{
Perform a try to submit the filled form
contains validation und writing the datas if validation is valid
}
\examples{
apply_submit(input, output, "orders_dir")
}
