% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrProtIdsFromFasta.R
\name{extrProtIdsFromFasta}
\alias{extrProtIdsFromFasta}
\title{tries to extract the protein IDs (located in fasta headers) from fasta file}
\usage{
extrProtIdsFromFasta(fasta_file, sub_pattern = NULL)
}
\arguments{
\item{fasta_file}{path to a fasta file}

\item{sub_pattern}{optionally, pattern to match the protein ID}
}
\value{
a character vector of protein IDs
}
\description{
tries to extract the protein IDs (located in fasta headers) from fasta file
}
\examples{
\dontrun{
extrProtIdsFromFasta("HUMAN_DB.fasta", ">([^ ]+)")
}
}
