% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMqparTemplate.R
\name{fillMqparTemplate}
\alias{fillMqparTemplate}
\title{fills a template mqpar string with the given infos}
\usage{
fillMqparTemplate(parameterGroups, fastaFiles, rawFiles = NULL,
  templateFile = "mqpar.tmp.xml", fixedFolder = "")
}
\arguments{
\item{fastaFiles}{vector of fasta files paths}

\item{rawFiles}{optionaly vector of raw files paths}

\item{templateFile}{template File path}

\item{paramterGroups}{list of parameterGroups settings (mode, enzymes)}
}
\value{
filled mqpar string
}
\description{
fills a template mqpar string with the given infos
}
\examples{
\dontrun{
fillMqparTemplate( list(list(mode=0, enzyme=c('LysC','GluC'))), c("human.fasta", "yeast.fasta"))
}
}
