% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilenamesFromChosenDBs.R
\name{getFilenamesFromChosenDBs}
\alias{getFilenamesFromChosenDBs}
\title{extracts databases file paths on the basis of selected databases in the form}
\usage{
getFilenamesFromChosenDBs(input, available_dbs)
}
\arguments{
\item{available_dbs}{a dataframe with the DBs informations}

\item{inputs}{isolated input list}
}
\value{
file paths corresponding to the choosen DBs in the form
}
\description{
extracts databases file paths on the basis of selected databases in the form
}
\examples{
getFilenamesFromChosenDBs(isolate(reactiveValuesToList(input)))
}
