% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilenamesFromDBnames.R
\name{getFilenamesFromDBnames}
\alias{getFilenamesFromDBnames}
\title{extracts databases file paths on the basis of selected databases in the form}
\usage{
getFilenamesFromDBnames(db_names)
}
\arguments{
\item{db_names}{names of fasta databases}
}
\value{
file paths corresponding to the databases names
}
\description{
extracts databases file paths on the basis of selected databases in the form
}
\examples{
\dontrun{
getFilenamesFromDBnames("Human")
}
}
