% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtIdRegexFromFileName.R
\name{getProtIdRegexFromFileName}
\alias{getProtIdRegexFromFileName}
\title{tries to determine an appropriate regex for Protein ID extraction from fasta headers
on the basis of file names}
\usage{
getProtIdRegexFromFileName(file_name, default_pattern = ">([^ ]+)")
}
\arguments{
\item{file_name}{one or more(vector) file names}

\item{default_pattern}{regex wich is used in case of no recognition}
}
\value{
a vector of regex
}
\description{
tries to determine an appropriate regex for Protein ID extraction from fasta headers
on the basis of file names
}
\examples{
getProtIdRegexFromFileName("Uniprot_Human_DB.fasta")
}
