% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markRequiredInput.R
\name{markRequiredInput}
\alias{markRequiredInput}
\title{Marks an input field as required
appends an asterix to the label und a message unter the label}
\usage{
markRequiredInput(inputId, message = "")
}
\arguments{
\item{inputId}{the id of an input field}

\item{message}{a messaged showed under the input field label}
}
\description{
Marks an input field as required
appends an asterix to the label und a message unter the label
}
\examples{
markRequiredInput("surename", "field is mandatory")
}
