% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeFastaSeq.R
\name{normalizeFastaSeq}
\alias{normalizeFastaSeq}
\title{replaces whitespaces in the headers
replaces whitespaces with empty string in other lines}
\usage{
normalizeFastaSeq(sequence)
}
\arguments{
\item{sequence}{string containing sequences in fasta format}
}
\value{
modificated fasta string
}
\description{
replaces whitespaces in the headers
replaces whitespaces with empty string in other lines
}
\examples{
normalizeFastaSeq(">Protein ID with spaces \\n KDFSHGZWVBSZWHA  ADFHDJF\\n")
}
