% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMqparTemplate.R
\name{preparePathForSub}
\alias{preparePathForSub}
\title{normalize and prepare one or more paths for substitution into a string template
e.g. for Max Quant template string}
\usage{
preparePathForSub(paths)
}
\arguments{
\item{paths}{paths to be prepared}
}
\value{
prepared path(s)
}
\description{
normalize and prepare one or more paths for substitution into a string template
e.g. for Max Quant template string
}
\examples{
\dontrun{
preparePathForSub("test_dir/fasta_file.fasta")
}
}
