% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAndPrintHtmlFile.R
\name{showAndPrintHtmlFile}
\alias{showAndPrintHtmlFile}
\title{opens a new browser window or tab where the html will be loaded
and tries to print it}
\usage{
showAndPrintHtmlFile(file, name_replacement = NULL)
}
\arguments{
\item{name_replacement}{optional, new name for the html file}

\item{path}{server path to a file}
}
\value{
html path relative to www directory
}
\description{
opens a new browser window or tab where the html will be loaded
and tries to print it
}
\examples{
\dontrun{
showAndPrintHtmlFile("form.html", "tmp")
}
}
