% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCompletedForm.R
\name{submitCompletedForm}
\alias{submitCompletedForm}
\title{gather all necessary inputs and uploaded files,
enriches them if necessary and puts them in the output dir}
\usage{
submitCompletedForm(allInputs, out_directory)
}
\arguments{
\item{allInputs}{all isolated inputs}

\item{out_directory}{directory where the outputs well be written to}
}
\value{
file path to the completed form html and the corresponding web path
}
\description{
gather all necessary inputs and uploaded files,
enriches them if necessary and puts them in the output dir
}
\examples{
\dontrun{
submitCompletedForm(isolate(reactiveValuesToList(input), "custormer_XY_date_AB")
}
}
