% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBarcodes.R
\name{summarizeBarcodes}
\alias{summarizeBarcodes}
\title{summarize the barcodes into one named list
names are the anzyme names, values are the barcodes respectively}
\usage{
summarizeBarcodes(allInputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\description{
summarize the barcodes into one named list
names are the anzyme names, values are the barcodes respectively
}
\examples{
\dontrun{
summarizeBarcodes(isolate(reactiveValuesToList(input)))
}
}
