% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateBarCodes.R
\name{validateBarCodes}
\alias{validateBarCodes}
\title{performs a validation check on the barcodes fields}
\usage{
validateBarCodes(inputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\value{
list of validation results
}
\description{
performs a validation check on the barcodes fields
}
\examples{
validateBarCodes(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
