% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataBases.R
\name{validateDataBases}
\alias{validateDataBases}
\title{performs a validation on databases input fields}
\usage{
validateDataBases(inputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\value{
list of validation results
}
\description{
performs a validation on databases input fields
}
\examples{
validateDataBases(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
