% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFastaFiles.R
\name{validateFastaFiles}
\alias{validateFastaFiles}
\title{validates whether the files are valid fasta files
uses read.fasta from seqinr to do that}
\usage{
validateFastaFiles(file_path)
}
\arguments{
\item{file_path}{fasta file paths}
}
\value{
logical, whether the file is a fasta file or NULL if file_path is NULL
}
\description{
validates whether the files are valid fasta files
uses read.fasta from seqinr to do that
}
\examples{
\dontrun{
validateFastaFiles("fasta_DB.fasta")
}
}
