% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFastaInputs.R
\name{validateFastaInputs}
\alias{validateFastaInputs}
\title{validates input fields related to uploaded fasta files}
\usage{
validateFastaInputs(inputs, input_ids = c("custom_species_db_file",
  "custom_background_db_file"))
}
\arguments{
\item{inputs}{isolated inputs}

\item{input_ids}{input fields ids}
}
\value{
list of validation results
}
\description{
validates input fields related to uploaded fasta files
}
\examples{
\dontrun{
validateFastaInputs(isolate(reactiveValuesToList(input)))
}
}
\seealso{
Validated
}
