% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateGeneralSettings.R
\name{validateGeneralSettings}
\alias{validateGeneralSettings}
\title{performs a validation on miscellaneous settings}
\usage{
validateGeneralSettings(inputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\value{
list of validation results
}
\description{
performs a validation on miscellaneous settings
}
\examples{
validateUserInformation(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
