% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateInputValues.R
\name{validateInputValues}
\alias{validateInputValues}
\title{performs a validation on the whole web-form}
\usage{
validateInputValues(inputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\value{
list of validation results
}
\description{
performs a validation on the whole web-form
}
\examples{
validateInputValues(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
