% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateUserInformation.R
\name{validateUserInformation}
\alias{validateUserInformation}
\title{performs a validation on user infos fields}
\usage{
validateUserInformation(inputs, user_inputs_info_ids)
}
\arguments{
\item{inputs}{isolated input list}

\item{user_inputs_info_ids}{list of input filed ids}
}
\value{
list of validation results
}
\description{
performs a validation on user infos fields
}
\examples{
validateUserInformation(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
