% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateBarCodes.R
\name{validateBarCodes}
\alias{validateBarCodes}
\title{performs a validation check on the barcodes fields}
\usage{
validateBarCodes(inputs, pattern = "^IMB_\\\\w+_\\\\d{3,5}$")
}
\arguments{
\item{inputs}{isolated input list}

\item{pattern}{regex pattern to be matched the barcodes}
}
\value{
list of validation results
}
\description{
performs a validation check on the barcodes fields
}
\examples{
validateBarCodes(isolate(reactiveValuesToList(input)))
}
\seealso{
Validated
}
