% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validated.R
\name{Validated}
\alias{Validated}
\title{creates a status object (named list)
comfort function}
\usage{
Validated(input_element_id, message = NULL, valid = FALSE)
}
\arguments{
\item{message}{message string}

\item{valid}{validation status}

\item{input_elemnt_id}{id of a input element}
}
\value{
anmed list containing the inputs
}
\description{
creates a status object (named list)
comfort function
}
\examples{
Validated("surename", "Mandatory field", FALSE )
}
