% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFastaSequence.R
\name{validateFastaSequences}
\alias{validateFastaSequences}
\title{validates a fasta sequence string}
\usage{
validateFastaSequences(seqs)
}
\arguments{
\item{sequence}{string containing one or more fasta sequences}
}
\value{
logical, whether the file is a fasta file or NULL if file_path is NULL
}
\description{
validates a fasta sequence string
}
\examples{
validateFastaSequences(seqs)
}
