% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateUserInformation.R
\name{validateUserInformation}
\alias{validateUserInformation}
\title{performs a validation on user infos fields}
\usage{
validateUserInformation(inputs)
}
\arguments{
\item{inputs}{isolated input list}
}
\value{
list of validation results
}
\description{
performs a validation on user infos fields
}
\examples{
\dontrun{
validateUserInformation(isolate(reactiveValuesToList(input)))
}
}
\seealso{
Validated
}
